ert_pinn_framework/requirements.txt
numpy>=1.24
PyYAML>=6.0
torch>=2.2
tqdm>=4.66
